+++ sdccc/src/main/java/com/draeger/medical/sdccc/configuration/EnabledTestConfig.java
@@ -65,6 +65,7 @@ public final class EnabledTestConfig {
     public static final String BICEPS_R5042 = BICEPS + "R5042";
     public static final String BICEPS_R5051 = BICEPS + "R5051";
     public static final String BICEPS_R5052 = BICEPS + "R5052";
+    public static final String BICEPS_R5053 = BICEPS + "R5053";
 
     public static final String BICEPS_547_0_0 = BICEPS + "5-4-7_0_0";
     public static final String BICEPS_547_1 = BICEPS + "5-4-7_1";