+++ sdccc/src/main/java/com/draeger/medical/sdccc/tests/biceps/invariant/InvariantMessageModelAnnexTest.java
@@ -35,7 +35,6 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.BeforeEach;
@@ -287,24 +286,21 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
     private void checkNestedDescriptors(final AbstractDescriptor descriptor) {
         final var handle = descriptor.getHandle();
         final String errorMsg = "%s with handle %s should not have nested descriptor %s";
-        if (descriptor instanceof AlertSystemDescriptor) {
+        if (descriptor instanceof AlertSystemDescriptor alertSystem) {
             // verify that the alarm system has no alarm signals or alarm conditions
-            final var alertSystem = (AlertSystemDescriptor) descriptor;
             assertTrue(
                     alertSystem.getAlertCondition().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, alertSystem.getAlertCondition()));
             assertTrue(
                     alertSystem.getAlertSignal().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, alertSystem.getAlertSignal()));
-        } else if (descriptor instanceof ChannelDescriptor) {
+        } else if (descriptor instanceof ChannelDescriptor channel) {
             // verify that the channel has no metrics
-            final var channel = (ChannelDescriptor) descriptor;
             assertTrue(
                     channel.getMetric().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, channel.getMetric()));
-        } else if (descriptor instanceof MdsDescriptor) {
+        } else if (descriptor instanceof MdsDescriptor mds) {
             // verify that the mds has no alert system, sco, system context, clock, batteries or vmds
-            final var mds = (MdsDescriptor) descriptor;
             assertNull(
                     mds.getAlertSystem(), String.format(errorMsg, descriptor.getClass(), handle, mds.getAlertSystem()));
             assertNull(mds.getSco(), String.format(errorMsg, descriptor.getClass(), handle, mds.getSco()));
@@ -316,16 +312,14 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
                     mds.getBattery().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, mds.getBattery()));
             assertTrue(mds.getVmd().isEmpty(), String.format(errorMsg, descriptor.getClass(), handle, mds.getVmd()));
-        } else if (descriptor instanceof ScoDescriptor) {
+        } else if (descriptor instanceof ScoDescriptor sco) {
             // verify that the sco has no operations
-            final var sco = (ScoDescriptor) descriptor;
             assertTrue(
                     sco.getOperation().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, sco.getOperation()));
-        } else if (descriptor instanceof SystemContextDescriptor) {
+        } else if (descriptor instanceof SystemContextDescriptor systemContext) {
             // verify that the system context has no patient and location context and no ensemble, operator
             // workflow and mean contexts
-            final var systemContext = (SystemContextDescriptor) descriptor;
             assertNull(
                     systemContext.getPatientContext(),
                     String.format(errorMsg, descriptor.getClass(), handle, systemContext.getPatientContext()));
@@ -344,9 +338,8 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
             assertTrue(
                     systemContext.getMeansContext().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, systemContext.getMeansContext()));
-        } else if (descriptor instanceof VmdDescriptor) {
+        } else if (descriptor instanceof VmdDescriptor vmd) {
             // verify that the vmd has no channels, vmd and sco
-            final var vmd = (VmdDescriptor) descriptor;
             assertTrue(
                     vmd.getChannel().isEmpty(),
                     String.format(errorMsg, descriptor.getClass(), handle, vmd.getChannel()));
@@ -376,7 +369,7 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
                     final var crtReportParts = reportOpt.orElseThrow().getReportPart().stream()
                             .filter(part ->
                                     ImpliedValueUtil.getModificationType(part).equals(DescriptionModificationType.CRT))
-                            .collect(Collectors.toList());
+                            .toList();
                     for (var part : crtReportParts) {
                         for (var state : part.getState()) {
                             acceptableSequenceSeen.incrementAndGet();
@@ -416,7 +409,7 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
 
             assertTestData(
                     acceptableSequenceSeen.get(),
-                    "No report parts with description modification type crt" + " seen during test run, test failed");
+                    "No report parts with description modification type crt seen during test run, test failed");
         }
     }
 
@@ -440,7 +433,7 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
                     final var uptReportParts = reportOpt.orElseThrow().getReportPart().stream()
                             .filter(part ->
                                     ImpliedValueUtil.getModificationType(part).equals(DescriptionModificationType.UPT))
-                            .collect(Collectors.toList());
+                            .toList();
                     for (var part : uptReportParts) {
                         acceptableSequenceSeen.incrementAndGet();
                         for (var state : part.getState()) {
@@ -483,7 +476,43 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
 
             assertTestData(
                     acceptableSequenceSeen.get(),
-                    "No report parts with description modification type upt" + " seen during test run, test failed");
+                    "No report parts with description modification type upt seen during test run, test failed");
+        }
+    }
+
+    @Test
+    @TestIdentifier(EnabledTestConfig.BICEPS_R5053)
+    @TestDescription("Retrieves every report part with modification type del from every description modification report"
+            + " seen during the test run and checks if the states are excluded from the message.")
+    @RequirePrecondition(simplePreconditions = {ConditionalPreconditions.DescriptionModificationDelPrecondition.class})
+    void testRequirementR5053() throws NoTestData, IOException {
+        try (final var messages =
+                messageStorage.getInboundMessagesByBodyType(Constants.MSG_DESCRIPTION_MODIFICATION_REPORT)) {
+            final var delReportsSeen = new AtomicInteger(0);
+
+            messages.getStream().forEach(messageContent -> {
+                try {
+                    final var soapMessage = marshalling.unmarshal(
+                            new ByteArrayInputStream(messageContent.getBody().getBytes(StandardCharsets.UTF_8)));
+                    final var reportOpt = soapUtil.getBody(soapMessage, DescriptionModificationReport.class);
+                    final var delReportParts = reportOpt.orElseThrow().getReportPart().stream()
+                            .filter(part ->
+                                    ImpliedValueUtil.getModificationType(part).equals(DescriptionModificationType.DEL))
+                            .toList();
+                    for (var part : delReportParts) {
+                        delReportsSeen.incrementAndGet();
+                        assertTrue(
+                                part.getState().isEmpty(),
+                                "State should not be part of the report with modification type delete.");
+                    }
+                } catch (MarshallingException e) {
+                    fail("Error unmarshalling MessageContent " + e);
+                }
+            });
+
+            assertTestData(
+                    delReportsSeen.get(),
+                    "No report parts with description modification type del seen during test run, test failed");
         }
     }
 
@@ -550,7 +579,7 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
             + " EpisodicComponentReport.")
     @RequirePrecondition(
             simplePreconditions = {ConditionalPreconditions.TriggerEpisodicComponentReportPrecondition.class})
-    void testRequirementC12() throws NoTestData, PreprocessingException, ReportProcessingException, IOException {
+    void testRequirementC12() throws NoTestData, IOException {
         final var mdibHistorian = mdibHistorianFactory.createMdibHistorian(
                 messageStorage, getInjector().getInstance(TestRunObserver.class));
 
@@ -607,7 +636,7 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
             + " EpisodicContextReport.")
     @RequirePrecondition(
             simplePreconditions = {ConditionalPreconditions.TriggerEpisodicContextReportPrecondition.class})
-    void testRequirementC13() throws NoTestData, PreprocessingException, ReportProcessingException, IOException {
+    void testRequirementC13() throws NoTestData, IOException {
         final var mdibHistorian = mdibHistorianFactory.createMdibHistorian(
                 messageStorage, getInjector().getInstance(TestRunObserver.class));
 
@@ -675,7 +704,7 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
             + " whether at least one child or attribute has changed for each AbstractMetricState contained in an"
             + " EpisodicMetricReport.")
     @RequirePrecondition(simplePreconditions = {ConditionalPreconditions.TriggerEpisodicMetricReportPrecondition.class})
-    void testRequirementC14() throws NoTestData, PreprocessingException, ReportProcessingException, IOException {
+    void testRequirementC14() throws NoTestData, IOException {
         final var mdibHistorian = mdibHistorianFactory.createMdibHistorian(
                 messageStorage, getInjector().getInstance(TestRunObserver.class));
 
@@ -693,11 +722,11 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
                     for (final Iterator<AbstractReport> iterator = reports.iterator(); iterator.hasNext(); ) {
                         final AbstractReport report = iterator.next();
 
-                        if (report instanceof EpisodicMetricReport) {
+                        if (report instanceof EpisodicMetricReport metricReport) {
                             acceptableSequenceSeen.incrementAndGet();
                             second = mdibHistorian.applyReportOnStorage(second, report);
 
-                            for (var reportPart : ((EpisodicMetricReport) report).getReportPart()) {
+                            for (var reportPart : metricReport.getReportPart()) {
                                 for (var metricState : reportPart.getMetricState()) {
                                     final var stateBeforeReport = first.getState(
                                             metricState.getDescriptorHandle(), AbstractMetricState.class);
@@ -752,11 +781,11 @@ public class InvariantMessageModelAnnexTest extends InjectorTestBase {
                             reportIterator.hasNext(); ) {
                         final AbstractReport report = reportIterator.next();
 
-                        if (report instanceof EpisodicOperationalStateReport) {
+                        if (report instanceof EpisodicOperationalStateReport operationalStateReport) {
                             acceptableSequenceSeen.incrementAndGet();
                             second = mdibHistorian.applyReportOnStorage(second, report);
 
-                            for (var reportPart : ((EpisodicOperationalStateReport) report).getReportPart()) {
+                            for (var reportPart : operationalStateReport.getReportPart()) {
                                 for (var operationState : reportPart.getOperationState()) {
                                     final var stateBeforeReport = first.getState(
                                             operationState.getDescriptorHandle(), AbstractOperationState.class);